Resubmission of the ICON ProgNa (alt) experiment
These fix an issue with too great moisture flux from the surface and the
resulting too high IWP and optical depth. Riming to graupel is also reduced by a
tunable parameter of the 2-moment microphysics scheme alpha_spacefilling set to
0.68 (slow riming), which was previously 0.01 (fast riming).

The meaning of the alt experiment is described in the alt NetCDF attribute. We
do not submit ProgNaNi due to the way INPs are implemented in our 2-moment
scheme configuration.

This commit also removes the ICON ProgNa and ProgNaNi experiments in sandbox
because they are affected by the abovementioned issue.